+++ src/main/java/at/rtr/rmbt/request/CellLocationRequest.java
@@ -18,7 +18,7 @@ public class CellLocationRequest {
     private final Long timeNs;
 
     @JsonProperty(value = "area_code")
-    private final Integer areaCode;
+    private final Long areaCode;
 
     @JsonProperty(value = "location_id")
     private final Long locationId;
